Configured webpack
