+++ .gitignore
@@ -5,7 +5,7 @@
 *.pdf
 *.psd
 *.sublime*
-build/
+build/js
 node_modules/
 npm-debug.log*
 Thumbs.db
