+++ source/js/main.js
@@ -1,6 +1,4 @@
 import './photo.js'
-import './map.js'
-import './form.js'
 import {getData} from './api.js';
 import {createMarkersAds} from './map.js';
 import {setFormSubmit} from './form.js';
