+++ source/js/util.js
@@ -1,0 +1,41 @@
+const ALERT_SHOW_TIME = 5000;
+
+const showAlert = (message) => {
+  const alertContainer = document.createElement('div');
+  alertContainer.style.zIndex = 500;
+  alertContainer.style.position = 'absolute';
+  alertContainer.style.left = 0;
+  alertContainer.style.top = '10%';
+  alertContainer.style.right = 0;
+  alertContainer.style.padding = '10px 3px';
+  alertContainer.style.fontSize = '30px';
+  alertContainer.style.textAlign = 'center';
+  alertContainer.style.backgroundColor = 'red';
+  
+  alertContainer.textContent = message;
+  
+  document.body.append(alertContainer);
+
+  setTimeout(() => {
+    alertContainer.remove();
+  }, ALERT_SHOW_TIME);
+}
+
+const isEscEvent = (evt) => {
+  return evt.key === 'Escape' || evt.key === 'Esc';
+};
+
+const isClickEvent = (evt) => {
+  return evt.type === 'click';
+};
+
+const debounce = (filterAds, timeout) => {
+  let timeFilter;
+  return () => {
+    clearTimeout(timeFilter);
+    timeFilter = setTimeout(() => filterAds.apply(this), timeout);
+  };
+};
+
+
+export {isEscEvent, isClickEvent, showAlert, debounce};
