+++ webpack.config.js
@@ -1,0 +1,10 @@
+const path = require('path');
+
+module.exports = {
+  entry: './source/js/main.js',
+  devtool: 'source-map',
+  output : {
+    filename: 'main.bundle.js',
+    path: path.resolve(__dirname, 'build/js'),
+  }
+};
